Break the components in small pieces
