+++ examples/react/package.json
@@ -13,20 +13,20 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@telnyx/webrtc": "2.1.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.6.2",
+    "@telnyx/webrtc": "2.1.0",
+    "styled-components": "^4.4.0",
     "@storybook/addon-actions": "^5.2.1",
     "@storybook/addon-console": "^1.2.1",
     "@storybook/addon-knobs": "^5.2.1",
     "@storybook/addon-links": "^5.2.1",
     "@storybook/addons": "^5.2.1",
-    "@storybook/react": "^5.2.1",
+    "@storybook/react": "^5.2.1"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.6.2",
     "babel-loader": "^8.0.6",
     "babel-plugin-react-docgen": "^3.1.0",
     "react-proptypes": "^1.0.0",
-    "storybook-addon-smart-knobs": "^5.0.0",
-    "styled-components": "^4.4.0"
+    "storybook-addon-smart-knobs": "^5.0.0"
   }
 }
